requirements.txt
certifi==2021.5.30
chardet==3.0.4
idna==2.9
requests==2.25.1
urllib3==1.26.6
python-telegram-bot==12.8